Make utils module Move config related functions to utils/config_file_helper.rs
