+++ src/cli/commands/setup.rs
@@ -6,11 +6,11 @@ use std::{
 
 use serde::Serialize;
 use clap::Args;
-use crate::blocking::config::config::find_config_dir;
+use crate::utils::config_file_helper;
 
 pub fn cmd_setup(setup_args: &SetupConfigArgs) -> Result<(), String> {
     // Use helper to get config directory and ensure it exists
-    let config_dir: PathBuf = find_config_dir().expect("Could not find config directory");
+    let config_dir: PathBuf = config_file_helper::find_config_dir().expect("Could not find config directory");
     fs::create_dir_all(&config_dir).expect("Could not create config directory");
 
     // Create hosts_backup file if it doesn't exist
