+++ src/lib.rs
@@ -1,16 +1,8 @@
 pub mod os_backend;
 
 pub mod blocking {
-    pub mod config {
-        pub mod config;
-    }
 
     pub mod websites;
-}
-
-pub mod ui {
-    pub mod blocking_message;
-    pub mod spinners;
 }
 
 pub mod config {}
@@ -23,4 +15,13 @@ pub mod cli {
         pub mod setup;
         pub mod start;
     }
+}
+
+pub mod ui {
+    pub mod blocking_message;
+    pub mod spinners;
+}
+
+pub mod utils {
+    pub mod config_file_helper;
 }
